Make GTK UI optional
