+++ ui/Cargo.toml
@@ -4,11 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-gtk4 = { version = "0.9", package = "gtk4", features = ["v4_10"] }
-glib = "0.18"
-gtk4-layer-shell = "0.5"
-gdk4 = { version = "0.9", features = ["v4_10"] }
+gtk4 = { version = "0.9", package = "gtk4", features = ["v4_10"], optional = true }
+glib = { version = "0.18", optional = true }
+gtk4-layer-shell = { version = "0.5", optional = true }
+gdk4 = { version = "0.9", features = ["v4_10"], optional = true }
 libc = "0.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["full"] }
+
+[features]
+default = []
+gui = ["gtk4", "glib", "gtk4-layer-shell", "gdk4"]
